pyproject.toml
[project]
name = "xhs-food-agent"
version = "1.0.0"
description = "小红书美食智能推荐Agent - 独立版"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # FastAPI and server
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sse-starlette>=1.8.0",
    
    # LangChain
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-core>=0.1.0",
    
    # HTTP clients
    "httpx>=0.25.0",
    "requests>=2.31.0",
    
    # Data validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    
    # Crypto for XHS API
    "pycryptodome>=3.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/xhs_food", "src/api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
